Port to x64; add heartbeat.

+++ serial_bridge.h
@@ -2,4 +2,4 @@
 
 extern void serial_bridge_init(struct config *);
 extern void serial_bridge_cleanup();
-extern void serial_write(char *, int);
+extern void serial_write(char *, size_t);
