+++ tcp_bridge.h
@@ -2,4 +2,4 @@
 
 extern int tcp_bridge_init(struct config *);
 extern void tcp_bridge_cleanup();
-extern void tcp_write(char *buffer, int length);
+extern void tcp_write(char *buffer, size_t length);
